+++ src/lib/render.ts
@@ -1,33 +1,29 @@
 import { convert } from 'html-to-text';
-import pretty from 'pretty';
-import type { ComponentProps, ComponentType, SvelteComponent } from 'svelte';
+import type { ComponentProps, Component, SvelteComponent } from 'svelte';
+import { render as renderServer } from 'svelte/server';
 
-export const render = <Component extends SvelteComponent>({
-	template,
-	props,
-	options
-}: {
-	template: ComponentType<Component>;
-	props?: ComponentProps<Component>;
-	options?: {
-		plainText?: boolean;
-		pretty?: boolean;
-	};
-}) => {
-	const { html } =
-		// @ts-ignore
-		template.render(props);
-	if (options?.plainText) {
-		return renderAsPlainText(html);
-	}
-	const doctype =
-		'<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">';
-	const markup = html;
-	const document = `${doctype}${markup}`;
-	if (options?.pretty) {
-		return pretty(document);
-	}
-	return document;
+export const render = async <
+  Comp extends SvelteComponent<any> | Component<any>,
+  Props extends ComponentProps<Comp> = ComponentProps<Comp>,
+>(
+  component: Comp,
+  props?: Props,
+) => {
+  const rendered = renderServer(component as any, {
+    props,
+  });
+
+  const doctype =
+    '<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">';
+
+  const html = `${doctype}${rendered.body}`;
+
+  const text = renderAsPlainText(rendered.body);
+
+  return {
+    html,
+    text,
+  };
 };
 
 const renderAsPlainText = (markup: string) => {